refactor: extract OutputFormat to utils::output_format
The connect-auth subcommands had their own OutputFormat enum plus
`emit_json_event` / `emit_json_object` helpers. Promote them to a
shared `crate::utils::output_format` module so init, config show,
install, build, and provision can all import the same type and use
the same NDJSON helpers without duplication. The auth module
re-exports the type under its old name for existing call sites.
