+++ src/utils/mod.rs
@@ -14,6 +14,7 @@ pub mod kernel_version;
 pub mod lockfile;
 pub mod nfs_server;
 pub mod output;
+pub mod output_format;
 pub mod pkcs11_devices;
 pub mod prerequisites;
 pub mod provision_result;
